Link between screens and backend
