+++ app/src/main/res/layout/activity_configs_consultant.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ConfigsConsultant">
+    tools:context=".activity.ConfigsConsultant">
 
     <ImageView
         android:id="@+id/imageView"
@@ -42,7 +42,7 @@
         android:background="@drawable/border_rounded"
         android:clipToOutline="true"
         android:scaleType="centerCrop"
-        android:src="@tools:sample/avatars"
+        android:src="@drawable/person_icon"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView" />
