+++ app/src/main/res/layout/activity_dashboard_consultant.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".DashboardConsultant">
+    tools:context=".activity.DashboardConsultant">
 <!--    <ScrollView-->
 <!--    xmlns:android="http://schemas.android.com/apk/res/android"-->
 <!--    xmlns:app="http://schemas.android.com/apk/res-auto"-->
@@ -377,10 +377,8 @@
             android:layout_gravity="center"
             app:indicatorColor="@color/white"
             app:indicatorSize="120dp"
-            app:max="100"
-            app:progress="10"
             app:trackColor="#0DA4A4"
-            app:trackThickness="12dp" />
+            app:trackThickness="12dp"/>
 
         <TextView
             android:id="@+id/tv_sales_goals"
@@ -410,8 +408,6 @@
             android:layout_height="match_parent"
             app:indicatorColor="#009688"
             app:indicatorSize="90dp"
-            app:max="100"
-            app:progress="50"
             app:trackColor="#0DA4A4"
             app:trackThickness="12dp" />
 
@@ -493,8 +489,6 @@
             android:layout_marginEnd="16dp"
             app:indicatorColor="#009688"
             app:indicatorSize="90dp"
-            app:max="100"
-            app:progress="50"
             app:trackColor="#0DA4A4"
             app:trackThickness="12dp" />
