+++ app/src/main/res/mipmap-anydpi-v26/download_icon_round.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
-    <background android:drawable="@drawable/download_icon_background"/>
+    <!-- <background android:drawable="@drawable/download_icon_background"/> -->
     <foreground android:drawable="@mipmap/download_icon_foreground"/>
 </adaptive-icon>
